[Improve] パケット用変数の型を一部最適化

+++ Assets/Users/Endo/Scripts/Network/PlayerData.cs
@@ -6,7 +6,7 @@ public class PlayerData
     public string Name;
     public string Uuid;
     public string Address;
-    public int    Port;
+    public ushort Port;
 
     public Vector3    Position;
     public Quaternion Rotation;
@@ -30,5 +30,5 @@ public static class SelfPlayerData
     public static string Name;
     public static string Uuid;
     public static string Address;
-    public static int    Port;
+    public static ushort Port;
 }
